Cleaned and added files
Cleaned up the program and stuck with large vector implementation due to lower bug susceptibility. Altered some of the comments for conversion to Linux-compliant program. Also wrote readme and makefile although I don't know if the makefile works. To test, change the argv[] lines in main.cpp and test the make file.