Works 100% on Kamek including for large doc set
Added .clear() to the words vector so the only memory limit is the number of words in the largest file. Basically means as long as no single file has millions of words, this program can handle any set of documents. It successfully handles the large doc set. I also changed the CLion commands to work for Linux using argv[] and .c_str(). It should work on your Linux machine too